requirements.txt
fastapi
uvicorn
pydantic-settings
openai
python-multipart
python-dotenv
gradio
faiss-cpu
sentence-transformers
numpy
